Set XSLT version `3.0` explicitly in `p:xslt`
Avoid that an XProc processor uses another way for determining the XSLT
version to use than checking the version of the actual stylesheet. From
https://spec.xproc.org/3.1/steps/#c.xslt.11:

> If the step specifies a version, then that version of XSLT must be
used to process the transformation. [...] If the step does not specify
a version, the implementation may use any version it has available and
may use any means to determine what version to use, including, but not
limited to, examining the version of the stylesheet.